refactor parquetSuites
